Move logic back into central location
