+++ packages/backends/package.json
@@ -10,7 +10,12 @@
   },
   "exports": {
     ".": "./dist/index.mjs",
-    "./package.json": "./package.json"
+    "./package.json": "./package.json",
+    "./introspection/loaders/cjs": "./dist/introspection/loaders/cjs.cjs",
+    "./introspection/loaders/esm": "./dist/introspection/loaders/esm.mjs",
+    "./introspection/loaders/hooks": "./dist/introspection/loaders/hooks.mjs",
+    "./introspection/loaders/rolldown-esm": "./dist/introspection/loaders/rolldown-esm.mjs",
+    "./introspection/loaders/rolldown-hooks": "./dist/introspection/loaders/rolldown-hooks.mjs"
   },
   "repository": {
     "type": "git",
@@ -27,20 +32,30 @@
     "dist"
   ],
   "dependencies": {
-    "@vercel/cervel": "workspace:*",
-    "@vercel/introspection": "workspace:*",
-    "fs-extra": "11.1.0"
+    "@vercel/build-utils": "workspace:*",
+    "@vercel/nft": "1.3.0",
+    "execa": "3.2.0",
+    "fs-extra": "11.1.0",
+    "oxc-transform": "0.111.0",
+    "path-to-regexp": "8.3.0",
+    "resolve.exports": "2.0.3",
+    "rolldown": "1.0.0-rc.1",
+    "srvx": "0.8.9",
+    "tsx": "4.21.0",
+    "zod": "3.22.4"
+  },
+  "peerDependencies": {
+    "typescript": "^4.0.0 || ^5.0.0"
   },
   "devDependencies": {
     "@types/express": "5.0.3",
     "@types/fs-extra": "11",
     "@types/jest": "27.5.1",
     "@types/node": "20.11.0",
-    "@vercel/build-utils": "workspace:*",
-    "execa": "3.2.0",
     "hono": "4.10.1",
     "jest-junit": "16.0.0",
     "tsdown": "0.16.3",
+    "typescript": "4.9.5",
     "vite": "^5.1.6",
     "vitest": "^2.0.1"
   },
