+++ packages/backends/src/introspection/hono.ts
