+++ packages/backends/src/introspection/loaders/esm.ts
@@ -1,0 +1,5 @@
+import { register } from 'node:module';
+// Disabling network blocking logic while introspection is opt-in
+// import './block-network.js';
+
+register(new URL('./hooks.mjs', import.meta.url), import.meta.url);
