+++ packages/backends/src/introspection/loaders/rolldown-esm.ts
@@ -1,0 +1,3 @@
+import { register } from 'node:module';
+
+register(new URL('./rolldown-hooks.mjs', import.meta.url), import.meta.url);
