+++ packages/backends/test/unit.test.ts
@@ -55,7 +55,7 @@ describe('successful builds', async () => {
   for (const fixtureName of fixtures) {
     // Windows is just too slow to build these fixtures
     const isNestFixture = fixtureName.includes('nest');
-    it.skipIf(process.platform === 'win32' && isNestFixture).only(
+    it.skipIf(process.platform === 'win32' && isNestFixture)(
       `builds ${fixtureName}`,
       async () => {
         // Copy entire fixture to work dir so no parent node_modules can interfere
