+++ packages/cervel/package.json
@@ -32,15 +32,7 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@vercel/build-utils": "workspace:*",
-    "@vercel/nft": "1.3.0",
-    "execa": "3.2.0",
-    "nf3": "0.3.6",
-    "oxc-transform": "0.111.0",
-    "resolve.exports": "2.0.3",
-    "rolldown": "1.0.0-rc.1",
-    "srvx": "0.8.9",
-    "tsx": "4.21.0"
+    "@vercel/backends": "workspace:*"
   },
   "peerDependencies": {
     "typescript": "^4.0.0 || ^5.0.0"
