+++ packages/introspection/src/loaders/cjs.ts
@@ -1,27 +1,3 @@
-import Module from 'module';
-import { handle as handleHono } from '../hono.js';
-import { handle as handleExpress } from '../express.js';
-// Disabling network blocking logic while introspection is opt-in
-// import './block-network.js';
-
-const originalRequire = Module.prototype.require;
-
-(Module.prototype.require as any) = function (
-  this: any,
-  id: string,
-  ...args: any[]
-) {
-  const result = originalRequire.apply(this, [id, ...args] as [string]);
-
-  if (id === 'express') {
-    return handleExpress(result);
-  }
-  if (id === 'hono') {
-    return {
-      ...result,
-      Hono: handleHono(result),
-    };
-  }
-
-  return result;
-};
+// Re-export from @vercel/backends
+// This file is kept for backwards compatibility
+import '@vercel/backends/introspection/loaders/cjs';
